requirements.txt
requests
pandas
SQLAlchemy
PyMySQL
python-dotenv